add casts to allow compilation with versions of ESP32 board packages

+++ src/LXWiFiSACN.cpp
@@ -313,7 +313,7 @@ uint16_t LXWiFiSACN::parse_dmp_layer( uint16_t size ) {
 			}
         	
         	// but if haven't heard from source a for three seconds...
-        	if ( abs(millis()-_last_packet_a) > 3000 ) {
+        	if ( abs((long)(millis()-_last_packet_a)) > 3000 ) {
         		// no more source a
         		if ( _packet_buffer[SACN_PRIORITY_OFFSET] > _priority_b ) {
         			// this packet takes over as source a if it has a higher priority than b
@@ -359,7 +359,7 @@ uint16_t LXWiFiSACN::parse_dmp_layer( uint16_t size ) {
            
            // if b exists and is expired, erase it
            if ( _dmx_slots_b ) {
-			   if ( abs(millis()-_last_packet_b) > 3000 ) {
+			   if ( abs((long)(millis()-_last_packet_b)) > 3000 ) {
 				  clearDMXSourceB();
 			   }
 		   }
